// src/types/settings.interface.ts
export default interface Settings {
	flags: boolean
	follow_list: boolean
	update_check: boolean
	cp_chart: boolean
	steam_avatar: boolean
	completions_by_tier: boolean
	country_top_100: boolean
	hover_info: boolean
	map_cover_image: boolean
	points_per_rank: boolean
	completions_bar_chart: boolean
	toasts: boolean
	user_ratings_table: boolean
	user_ratings: boolean
	user_effects: boolean
}